Use a consistent type for code size
If we just read it as a size_t* we can read additional garbage so
define it as a size_t in the first place.

+++ vm/gc/code_manager.hpp
@@ -56,7 +56,7 @@ namespace rubinius {
     int total_allocated_;
     int total_freed_;
 
-    int bytes_used_;
+    size_t bytes_used_;
 
   public:
     int freed_resources() {
@@ -75,7 +75,7 @@ namespace rubinius {
       return shared_;
     }
 
-    int& size() {
+    size_t& size() {
       return bytes_used_;
     }
